ch21/my-agent-project/pyproject.toml
[project]
name = "my-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.19",
    "langgraph-sdk>=0.3.12",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
]
